Initial move of configuration service to the core
