+++ exonum/Cargo.toml
@@ -56,6 +56,7 @@ protobuf = { version = "2.2.0", features = ["with-serde"] }
 ctrlc = "3.1.1"
 rpassword = "2.1.0"
 zeroize = "0.5.2"
+lazy_static = "1.0.1"
 
 exonum_rocksdb = "0.7.4"
 exonum_sodiumoxide = { version = "0.0.20", optional = true }
@@ -65,7 +66,6 @@ exonum-derive = { version = "0.10.0", path = "../components/derive" }
 [dev-dependencies]
 pretty_assertions = "0.5.1"
 tempdir = "0.3.7"
-lazy_static = "1.0.1"
 proptest = "0.8.7"
 modifier = "0.1.0"
 criterion = "0.2.8"
