+++ exonum/src/runtime/configuration/api.rs
@@ -1,4 +1,4 @@
-// Copyright 2018 The Exonum Team
+// Copyright 2019 The Exonum Team
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use exonum::{
+use crate::{
     api::{self, ServiceApiBuilder, ServiceApiState},
     blockchain::{Schema as CoreSchema, StoredConfiguration},
     crypto::{CryptoHash, Hash},
