+++ exonum/src/runtime/configuration/doc/response-samples.md
