+++ exonum/src/runtime/configuration/tests/api.rs
@@ -1,4 +1,4 @@
-// Copyright 2018 The Exonum Team
+// Copyright 2019 The Exonum Team
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
 
 // spell-checker:ignore postpropose, postvote
 
-use exonum::{
+use crate::{
     blockchain::{Schema, StoredConfiguration},
     crypto::{CryptoHash, Hash},
     helpers::{Height, ValidatorId},
@@ -25,11 +25,11 @@ use super::{
     new_tx_config_propose, new_tx_config_vote, new_tx_config_vote_against, ConfigurationSchema,
     ConfigurationTestKit,
 };
-use crate::api::{
+use super::super::api::{
     ConfigHashInfo, ConfigInfo, FilterQuery, HashQuery, ProposeHashInfo, ProposeResponse,
     VoteResponse, VotesInfo,
 };
-use crate::SERVICE_NAME;
+use super::super::SERVICE_NAME;
 
 trait ConfigurationApiTest {
     fn actual_config(&self) -> ConfigHashInfo;
