+++ exonum/src/runtime/configuration/tests/mod.rs
@@ -1,4 +1,4 @@
-// Copyright 2018 The Exonum Team
+// Copyright 2019 The Exonum Team
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use exonum::{
+use crate::{
     blockchain::{Schema, StoredConfiguration},
     crypto::{hash, CryptoHash, Hash, HASH_SIZE},
     helpers::{Height, ValidatorId},
@@ -23,7 +23,7 @@ use exonum_testkit::{TestKit, TestKitBuilder, TestNode};
 
 use std::str;
 
-use crate::{
+use super::{
     config::ConfigurationServiceConfig, ConfigurationTransactions, Propose,
     Schema as ConfigurationSchema, Service as ConfigurationService, Vote, VoteAgainst,
     VotingDecision, SERVICE_NAME,
